task 13: Delete states

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -1,41 +1,33 @@
 #!/usr/bin/python3
 """
-This script deletes all State objects with a name containing the letter 'a' \
+This script deletes all State objects
+with a name containing the letter `a`
 from the database `hbtn_0e_6_usa`.
 """
 
-import sys
+from sys import argv
+from model_state import State, Base
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
-from model_state import Base, State
 
 if __name__ == "__main__":
-    if len(sys.argv) != 4:
-        print("Usage: ./13-model_state_delete_a.py <mysql username> \
-<mysql password> <database name>")
-        sys.exit(1)
+    """
+    Deletes State objects on the database.
+    """
 
-    username = sys.argv[1]
-    password = sys.argv[2]
-    dbname = sys.argv[3]
+    db_url = "mysql+mysqldb://{}:{}@localhost:3306/{}".format(
+        argv[1], argv[2], argv[3])
 
-    engine = create_engine(f'mysql+mysqldb://{username}:{password}\
-@localhost:3306/{dbname}', pool_pre_ping=True)
-
-    # Create a configured "Session" class
+    engine = create_engine(db_url)
     Session = sessionmaker(bind=engine)
 
     session = Session()
 
-    try:
-        states_to_delete = session.query(State).\
-            filter(State.name.ilike('%a%')).all()
-
-        for state in states_to_delete:
+    states = session.query(State).filter(State.name.contains('a'))
+    if states is not None:
+        for state in states:
             session.delete(state)
 
-        session.commit()
-    except Exception as e:
-        print(f"Error: {e}")
-    finally:
-        session.close()
+    session.commit()
+
+    session.close()
